pyproject.toml
[project]
name = "langgraph-mcp-agent"
version = "0.1.0"
authors = [{ name = "huy.d@hotmail.com" }]
description = "LangGraph MCP Agent with MLOps tooling for Databricks"
readme = "README.md"
requires-python = ">=3.10,<=3.13"
dependencies = [
    # LangGraph MCP Agent core dependencies
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain>=0.3.0",
    "databricks-mcp>=0.1.0",
    "databricks-langchain>=0.1.0",
    "databricks-agents>=0.1.0",
    "mlflow-skinny[databricks]>=2.0.0",
    "databricks-sdk>=0.29.0",
    "nest-asyncio>=1.5.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "mcp>=1.0.0",
    # Azure dependencies for Unity Catalog
    "azure-identity>=1.12.0",
    "azure-storage-blob>=12.0.0",
    "azure-core>=1.26.0",
    # CLI dependencies
    "click>=8.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "databricks-dlt",
    "databricks-connect>=15.4,<15.5",
    "databricks-bundles==0.276.0",
]

[project.scripts]
langgraph-agent = "langgraph_agent.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/langgraph_agent"]

[tool.black]
line-length = 125
